refactor(web): extract JoyCaption prompt corpus into src/training/ (sc-4199)
TrainingStudio.jsx (2.2k lines) embedded ~130 lines of JoyCaption captioner
prompt templates plus the pure prompt-assembly logic inline, mixed with the
dataset editor and config form — slow to navigate, and pure logic buried where
it can't be tested.

Move the JoyCaption corpus (model id, type/length/extra-option lists, the
template map, default caption settings) and the pure buildJoyCaptionPrompt()
into a new src/training/joyCaptionPrompts.js module, imported by TrainingStudio.
TrainingStudio drops from 2263 to 2118 lines.

Adds src/training/joyCaptionPrompts.test.js covering buildJoyCaptionPrompt's
template-index selection (any/numeric/length), extra-option joining, and
{name}/{length}/{word_count} substitution — logic that was previously untestable.

Full web suite 383 passed, lint 0 errors, build clean.

Partial delivery of F-WEB-20: the remaining pure config/dataset helper modules
and the dataset-editor / configure-job panel component split are tracked as
follow-on work (story kept In Progress).
